draft the types needed for UCF
