+++ Sources/Codelinks/Codelink.Base.swift
@@ -1,0 +1,9 @@
+extension CodelinkV4
+{
+    @frozen public
+    enum Base:Equatable, Hashable, Sendable
+    {
+        case qualified
+        case relative
+    }
+}
