+++ Sources/Codelinks/V3/CodelinkV3.Filter.Subscript.swift
@@ -1,4 +1,4 @@
-extension Codelink.Filter
+extension CodelinkV3.Filter
 {
     @frozen public
     enum Subscript:Equatable, Hashable, Sendable
