+++ Sources/Codelinks/V3/CodelinkV3.Path.Format.swift
@@ -1,4 +1,4 @@
-extension Codelink.Path
+extension CodelinkV3.Path
 {
     enum Format:Equatable, Hashable, Sendable
     {
